Split Vercel project and domain helpers into get, add, verify, config
List stays with get. Check maps config nulls through toResult.
