+++ src/vercel/instance.ts
@@ -6,3 +6,9 @@ export function getVercel(): Vercel {
   const vercel = new Vercel({ bearerToken });
   return vercel;
 }
+
+export function getProject() {
+  const idOrName = getEnvVar("MEMOS_VERCEL_PROJECT_ID");
+  const vercel = getVercel();
+  return { idOrName, vercel };
+}
